NN-2203 pave the way for redis
