+++ package.json
@@ -20,7 +20,7 @@
     "build:hmpps": "cross-env CLIENT=hmpps NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p",
     "build:clean": "npm run test:clean && rimraf ./build",
     "build:dll": "node ./internals/scripts/dependencies.js",
-    "start": "cross-env CLIENT=hmpps NODE_ENV=development node server | bunyan -o short",
+    "start": "cross-env CLIENT=hmpps NODE_ENV=development node server",
     "start-feature": "cross-env CLIENT=hmpps NODE_ENV=development PORT=3007 USE_OF_FORCE_URL=http://use-of-force PATHFINDER_URL=http://pathfinder USE_OF_FORCE_PRISONS=LEI OAUTH_ENDPOINT_URL=http://localhost:19090/auth/ API_ENDPOINT_URL=http://localhost:18080/ KEYWORKER_API_URL=http://localhost:18081/ PRISON_STAFF_HUB_UI_URL=http://localhost:18082/ CASENOTES_API_URL=http://localhost:18083 node server | bunyan -o short",
     "start:tunnel": "cross-env NODE_ENV=development ENABLE_TUNNEL=true node server",
     "start:production": "npm run test && npm run build && npm run start:prod",
@@ -77,7 +77,6 @@
     "body-parser": "^1.17.2",
     "bootstrap": "4.0.0-alpha.6",
     "bunyan": "^1.8.12",
-    "bunyan-middleware": "^0.8.0",
     "chalk": "1.1.3",
     "cheerio": "^1.0.0-rc.2",
     "circular-dependency-plugin": "2.0.0",
